third_party/blink/web_tests/platform/generic/fast/dom/shadow/shadowhost-keyframes-expected.txt

Test if @keyframes works for shadow host when using :host. i.e. crbug.com/332577

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getColor(document.getElementById('host')) is "rgb(0, 128, 0)"
